% Auto-generated: do not edit by hand
\name{networkVis}

\alias{networkVis}

\title{NetworkVis component}

\description{

}

\usage{
networkVis(id=NULL, getEdges=NULL, getNetwork=NULL, getNodes=NULL,
graph=NULL, label=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{getEdges}{}

\item{getNetwork}{}

\item{getNodes}{}

\item{graph}{Named list. }

\item{label}{Named list. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

