% Auto-generated: do not edit by hand
\name{networkVis}

\alias{networkVis}

\title{NetworkVis component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
networkVis(id=NULL, edges=NULL, labelPosition=NULL, labelText=NULL,
labelVisible=NULL, nodes=NULL, selectedEdgeId=NULL,
selectedNodeId=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{edges}{List of named lists. }

\item{labelPosition}{Named list. }

\item{labelText}{Character. }

\item{labelVisible}{Logical. }

\item{nodes}{List of named lists. }

\item{selectedEdgeId}{Logical | numeric | character | named list | unnamed list. }

\item{selectedNodeId}{Logical | numeric | character | named list | unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

